fix: 공고 목록 타입 분리(#6)
