+++ src/types/jobPosts.ts
@@ -1,0 +1,11 @@
+export interface JobPost {
+  id: number
+  title: string
+  viewCount: number
+  commentCount: number
+  memberLimit: number
+  deadline: string
+  courses: string[]
+  tags: string[]
+  image: string
+}
